requirements.txt
# Trading & Data
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
ccxt>=4.0.0
ta>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0

# Backtesting
backtesting>=0.3.3

# Web3 / ERC-8004
web3>=6.0.0
eth-account>=0.10.0

# Dashboard
flask>=3.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0

# Utilities
schedule>=1.2.0
colorama>=0.4.6
